MERGE: 19541:19983 https://svn.centricsuite.com/webapp/branches/branch-41 includes 18995:19423 https://svn.centricsuite.com/webapp/branches/f-20070206-daffodildb NEW: DaffodilDB source upgrade scripts

+++ src/sql/daffodildb/new_website.sql
@@ -190,7 +190,7 @@ CREATE TABLE portfolio_category (
   parent_category_id INTEGER REFERENCES portfolio_category(category_id),
   enabled BOOLEAN DEFAULT true NOT NULL,
   entered TIMESTAMP DEFAULT CURRENT_TIMESTAMP NOT NULL,
-  enteredby NUMBER(10,0) REFERENCES "access"(user_id) NOT NULL,
+  enteredby INTEGER REFERENCES "access"(user_id) NOT NULL,
   modified TIMESTAMP DEFAULT CURRENT_TIMESTAMP NOT NULL,
   modifiedby INTEGER REFERENCES "access"(user_id) NOT NULL,
   PRIMARY KEY (CATEGORY_ID)
